apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "internatlas-api"
version = "0.1.0"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.30",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "psycopg2-binary>=2.9",
  "alembic>=1.13",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "bcrypt==4.0.1",
  "python-multipart>=0.0.9",
  "pgvector>=0.3",
  "httpx>=0.27",
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "aiosqlite>=0.20",
  "email-validator>=2.2",
  "google-genai",
  "beautifulsoup4"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*", "tests*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
